Added angular spread to the player charged fire weapon
Added game.charge_spread; the full spread angle in degrees. Hardwired to
3 for now.

The first argument of KOBO_myship::shot_single() is now floating point.

+++ src/myship.h
@@ -68,7 +68,7 @@ class KOBO_myship
 	// For the gfxengine connection
 	static cs_obj_t *object;
 
-	static void shot_single(int dir, int loffset, int hoffset);
+	static void shot_single(float dir, int loffset, int hoffset);
 	static void apply_position();
 	static void explode();
 	static void fire_control();
